+++ source/Burntime.Remaster/Scenes/OptionsSettingsPage.cs
@@ -14,7 +14,6 @@ internal class OptionsSettingsPage : Container
     readonly Button _musicToggle;
     readonly Button _newgfxToggle;
     readonly Button _fullscreenToggle;
-    readonly Button _scalingToggle;
     readonly Button _controllerToggle;
     readonly Button _languageToggle;
     readonly Button[] _buttons;
@@ -31,7 +30,7 @@ internal class OptionsSettingsPage : Container
             Font = _fonts.Green,
             HoverFont = _fonts.Orange,
             DisabledFont = _fonts.Disabled,
-            Position = new Vector2(38, 88),
+            Position = new Vector2(38, 78),
             IsTextOnly = true,
             IsEnabled = !BurntimeClassic.Instance.DisableMusic
         };
@@ -48,22 +47,6 @@ internal class OptionsSettingsPage : Container
             Font = _fonts.Green,
             HoverFont = _fonts.Orange,
             DisabledFont = _fonts.Disabled,
-            Position = new Vector2(38, 78),
-            IsTextOnly = true
-        };
-        Windows += _scalingToggle = new Button(app, () =>
-        {
-            OutputFiltering previous = app.Engine.OutputFiltering;
-            app.Engine.OutputFiltering = NextOutputFiltering(app.Engine);
-            if (!app.IsNewGfx &&
-                (previous == OutputFiltering.Xbr2 ||
-                 app.Engine.OutputFiltering == OutputFiltering.Xbr2))
-                app.Engine.ReloadGraphics();
-        })
-        {
-            Font = _fonts.Green,
-            HoverFont = _fonts.Orange,
-            DisabledFont = _fonts.Disabled,
             Position = new Vector2(38, 68),
             IsTextOnly = true
         };
@@ -73,7 +56,7 @@ internal class OptionsSettingsPage : Container
             Font = _fonts.Green,
             HoverFont = _fonts.Orange,
             DisabledFont = _fonts.Disabled,
-            Position = new Vector2(38, 108),
+            Position = new Vector2(38, 98),
             IsTextOnly = true
         };
         Windows += _controllerToggle = new Button(app,
@@ -82,7 +65,7 @@ internal class OptionsSettingsPage : Container
             Font = _fonts.Green,
             HoverFont = _fonts.Orange,
             DisabledFont = _fonts.Disabled,
-            Position = new Vector2(38, 98),
+            Position = new Vector2(38, 88),
             IsTextOnly = true
         };
         if (!app.Engine.SupportsFullscreenToggle)
@@ -93,14 +76,14 @@ internal class OptionsSettingsPage : Container
         Windows += _hintText = new Button(app)
         {
             Font = _fonts.Blue,
-            Position = new Vector2(40, 122),
+            Position = new Vector2(40, 112),
             Size = new Vector2(120, 10),
             TextHorizontalAlign = Platform.Graphics.TextAlignment.Center
         };
 
         _buttons = new[]
         {
-            _newgfxToggle, _scalingToggle, _fullscreenToggle, _musicToggle,
+            _newgfxToggle, _fullscreenToggle, _musicToggle,
             _controllerToggle, _languageToggle
         };
     }
@@ -183,13 +166,6 @@ internal class OptionsSettingsPage : Container
             _ => "@burn?424",
         };
         _fullscreenToggle.Text = app.Engine.IsFullscreen ? "@newburn?19" : "@newburn?20";
-        _scalingToggle.Text = app.Engine.OutputFiltering switch
-        {
-            OutputFiltering.NearestPoint => "Scaling POINT",
-            OutputFiltering.Linear => "Scaling LINEAR",
-            OutputFiltering.Xbr2 => "@newburn?56",
-            _ => "@newburn?55"
-        };
         _controllerToggle.Text = app.Engine.ControllerGlyphMode switch
         {
             ControllerGlyphMode.Xbox => "@controller?1",
@@ -225,20 +201,5 @@ internal class OptionsSettingsPage : Container
         }
 
         base.OnUpdate(elapsed);
-    }
-
-    static OutputFiltering NextOutputFiltering(IEngine engine)
-    {
-        if (engine.ForceNearestPointOutputFiltering)
-            return OutputFiltering.NearestPoint;
-        if (engine.ForceLinearOutputFiltering)
-            return OutputFiltering.Linear;
-        if (engine.DisableShaders || !engine.SupportsSharpBilinearShader)
-            return OutputFiltering.SharpBilinear;
-        if (!engine.SupportsXbr2Shader)
-            return OutputFiltering.SharpBilinearShader;
-        return engine.OutputFiltering == OutputFiltering.Xbr2
-            ? OutputFiltering.SharpBilinearShader
-            : OutputFiltering.Xbr2;
     }
 }